ivy 꺼랑 합침
