--- src/components/page/LoginPage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-
-function LoginPage(props) {
-return (
-    <p>로그인 페이지입니다. </p>
-)
-
-}
-
-export default LoginPage;
